internal/charmstore: factor out addentity tests
