+++ internal/charmstore/store_test.go
@@ -25,7 +25,6 @@ import (
 	"gopkg.in/errgo.v1"
 	"gopkg.in/juju/charm.v6-unstable"
 	"gopkg.in/juju/charmrepo.v2-unstable/csclient/params"
-	"gopkg.in/macaroon-bakery.v1/bakery"
 	"gopkg.in/mgo.v2/bson"
 	"gopkg.in/natefinch/lumberjack.v2"
 
@@ -38,243 +37,10 @@ import (
 )
 
 type StoreSuite struct {
-	storetesting.IsolatedMgoESSuite
-	index string
+	commonSuite
 }
 
 var _ = gc.Suite(&StoreSuite{})
-
-func (s *StoreSuite) checkAddCharm(c *gc.C, ch charm.Charm, addToES bool, url *router.ResolvedURL) {
-	var es *elasticsearch.Database
-	if addToES {
-		es = s.ES
-	}
-	store := s.newStore(c, true)
-	defer store.Close()
-
-	// Add the charm to the store.
-	beforeAdding := time.Now()
-	err := store.AddCharmWithArchive(url, ch)
-	c.Assert(err, gc.IsNil)
-	afterAdding := time.Now()
-
-	var doc *mongodoc.Entity
-	err = store.DB.Entities().FindId(&url.URL).One(&doc)
-	c.Assert(err, gc.IsNil)
-
-	// Ensure the document was indexed in ElasticSearch, if an ES database was provided.
-	if es != nil {
-		var result SearchDoc
-		id := store.ES.getID(doc.URL)
-		err = store.ES.GetDocument(s.TestIndex, typeName, id, &result)
-		c.Assert(err, gc.IsNil)
-		exists, err := store.ES.HasDocument(s.TestIndex, typeName, id)
-		c.Assert(err, gc.IsNil)
-		c.Assert(exists, gc.Equals, true)
-		if purl := url.DocPromulgatedURL(); purl != nil {
-			c.Assert(result.PromulgatedURL, jc.DeepEquals, purl)
-		}
-	}
-	// The entity doc has been correctly added to the mongo collection.
-	size, hash, hash256 := getSizeAndHashes(ch)
-	sort.Strings(doc.CharmProvidedInterfaces)
-	sort.Strings(doc.CharmRequiredInterfaces)
-
-	// Check the upload time and then reset it to its zero value
-	// so that we can test the deterministic parts later.
-	c.Assert(doc.UploadTime, jc.TimeBetween(beforeAdding, afterAdding))
-
-	doc.UploadTime = time.Time{}
-
-	blobName := doc.BlobName
-	c.Assert(blobName, gc.Matches, "[0-9a-z]+")
-	doc.BlobName = ""
-
-	c.Assert(doc, jc.DeepEquals, denormalizedEntity(&mongodoc.Entity{
-		URL:                     &url.URL,
-		BlobHash:                hash,
-		BlobHash256:             hash256,
-		Size:                    size,
-		CharmMeta:               ch.Meta(),
-		CharmActions:            ch.Actions(),
-		CharmConfig:             ch.Config(),
-		CharmProvidedInterfaces: []string{"http", "logging", "monitoring"},
-		CharmRequiredInterfaces: []string{"mysql", "varnish"},
-		PromulgatedURL:          url.DocPromulgatedURL(),
-		SupportedSeries:         ch.Meta().Series,
-		Development:             url.Development,
-	}))
-
-	// The charm archive has been properly added to the blob store.
-	r, obtainedSize, err := store.BlobStore.Open(blobName)
-	c.Assert(err, gc.IsNil)
-	defer r.Close()
-	c.Assert(obtainedSize, gc.Equals, size)
-	data, err := ioutil.ReadAll(r)
-	c.Assert(err, gc.IsNil)
-	charmArchive, err := charm.ReadCharmArchiveBytes(data)
-	c.Assert(err, gc.IsNil)
-	c.Assert(charmArchive.Meta(), jc.DeepEquals, ch.Meta())
-	c.Assert(charmArchive.Config(), jc.DeepEquals, ch.Config())
-	c.Assert(charmArchive.Actions(), jc.DeepEquals, ch.Actions())
-	c.Assert(charmArchive.Revision(), jc.DeepEquals, ch.Revision())
-
-	// Check that the base entity has been properly created.
-	assertBaseEntity(c, store, mongodoc.BaseURL(&url.URL), url.PromulgatedRevision != -1)
-
-	// Try inserting the charm again - it should fail because the charm is
-	// already there.
-	err = store.AddCharmWithArchive(url, ch)
-	c.Assert(errgo.Cause(err), gc.Equals, params.ErrDuplicateUpload)
-}
-
-// addRequiredCharms adds any charms required by the given
-// bundle that are not already in the store.
-func (s *StoreSuite) addRequiredCharms(c *gc.C, bundle charm.Bundle) {
-	store := s.newStore(c, true)
-	defer store.Close()
-	for _, svc := range bundle.Data().Services {
-		u := charm.MustParseURL(svc.Charm)
-		if _, err := store.FindBestEntity(u, nil); err == nil {
-			continue
-		}
-		if u.Revision == -1 {
-			u.Revision = 0
-		}
-		var rurl router.ResolvedURL
-		rurl.URL = *u
-		ch := storetesting.Charms.CharmDir(u.Name)
-		if len(ch.Meta().Series) == 0 && u.Series == "" {
-			rurl.URL.Series = "trusty"
-		}
-		if u.User == "" {
-			rurl.URL.User = "charmers"
-			rurl.PromulgatedRevision = rurl.URL.Revision
-		} else {
-			rurl.PromulgatedRevision = -1
-		}
-		err := store.AddCharmWithArchive(&rurl, ch)
-		c.Assert(err, gc.IsNil)
-	}
-}
-
-func (s *StoreSuite) checkAddBundle(c *gc.C, bundle charm.Bundle, addToES bool, url *router.ResolvedURL) {
-	var es *elasticsearch.Database
-
-	if addToES {
-		es = s.ES
-	}
-	store := s.newStore(c, true)
-	defer store.Close()
-
-	// Add the bundle to the store.
-	beforeAdding := time.Now()
-	s.addRequiredCharms(c, bundle)
-	err := store.AddBundleWithArchive(url, bundle)
-	c.Assert(err, gc.IsNil)
-	afterAdding := time.Now()
-
-	var doc *mongodoc.Entity
-	err = store.DB.Entities().FindId(&url.URL).One(&doc)
-	c.Assert(err, gc.IsNil)
-	sort.Sort(orderedURLs(doc.BundleCharms))
-
-	// Ensure the document was indexed in ElasticSearch, if an ES database was provided.
-	if es != nil {
-		var result SearchDoc
-		id := store.ES.getID(doc.URL)
-		err = store.ES.GetDocument(s.TestIndex, typeName, id, &result)
-		c.Assert(err, gc.IsNil)
-		exists, err := store.ES.HasDocument(s.TestIndex, typeName, id)
-		c.Assert(err, gc.IsNil)
-		c.Assert(exists, gc.Equals, true)
-		if purl := url.PromulgatedURL(); purl != nil {
-			c.Assert(result.PromulgatedURL, jc.DeepEquals, purl)
-		}
-	}
-
-	// Check the upload time and then reset it to its zero value
-	// so that we can test the deterministic parts later.
-	c.Assert(doc.UploadTime, jc.TimeBetween(beforeAdding, afterAdding))
-	doc.UploadTime = time.Time{}
-
-	// The blob name is random, but we check that it's
-	// in the correct format, and non-empty.
-	blobName := doc.BlobName
-	c.Assert(blobName, gc.Matches, "[0-9a-z]+")
-	doc.BlobName = ""
-
-	// The entity doc has been correctly added to the mongo collection.
-	size, hash, hash256 := getSizeAndHashes(bundle)
-	c.Assert(doc, jc.DeepEquals, denormalizedEntity(&mongodoc.Entity{
-		URL:          &url.URL,
-		BlobHash:     hash,
-		BlobHash256:  hash256,
-		Size:         size,
-		BundleData:   bundle.Data(),
-		BundleReadMe: bundle.ReadMe(),
-		BundleCharms: []*charm.URL{
-			charm.MustParseURL("mysql"),
-			charm.MustParseURL("wordpress"),
-		},
-		BundleMachineCount: newInt(2),
-		BundleUnitCount:    newInt(2),
-		PromulgatedURL:     url.DocPromulgatedURL(),
-		Development:        url.Development,
-	}))
-
-	// The bundle archive has been properly added to the blob store.
-	r, obtainedSize, err := store.BlobStore.Open(blobName)
-	c.Assert(err, gc.IsNil)
-	defer r.Close()
-	c.Assert(obtainedSize, gc.Equals, size)
-	data, err := ioutil.ReadAll(r)
-	c.Assert(err, gc.IsNil)
-	bundleArchive, err := charm.ReadBundleArchiveBytes(data)
-	c.Assert(err, gc.IsNil)
-	c.Assert(bundleArchive.Data(), jc.DeepEquals, bundle.Data())
-	c.Assert(bundleArchive.ReadMe(), jc.DeepEquals, bundle.ReadMe())
-
-	// Check that the base entity has been properly created.
-	assertBaseEntity(c, store, mongodoc.BaseURL(&url.URL), url.PromulgatedRevision != -1)
-
-	// Try inserting the bundle again - it should fail because the bundle is
-	// already there.
-	err = store.AddBundleWithArchive(url, bundle)
-	c.Assert(errgo.Cause(err), gc.Equals, params.ErrDuplicateUpload)
-}
-
-func assertBaseEntity(c *gc.C, store *Store, url *charm.URL, promulgated bool) {
-	baseEntity, err := store.FindBaseEntity(url, nil)
-	c.Assert(err, gc.IsNil)
-	expectACLs := mongodoc.ACL{
-		Read:  []string{url.User},
-		Write: []string{url.User},
-	}
-	c.Assert(baseEntity, jc.DeepEquals, &mongodoc.BaseEntity{
-		URL:             url,
-		User:            url.User,
-		Name:            url.Name,
-		Public:          false,
-		ACLs:            expectACLs,
-		DevelopmentACLs: expectACLs,
-		Promulgated:     mongodoc.IntBool(promulgated),
-	})
-}
-
-type orderedURLs []*charm.URL
-
-func (o orderedURLs) Less(i, j int) bool {
-	return o[i].String() < o[j].String()
-}
-
-func (o orderedURLs) Swap(i, j int) {
-	o[i], o[j] = o[j], o[i]
-}
-
-func (o orderedURLs) Len() int {
-	return len(o)
-}
 
 var urlFindingTests = []struct {
 	inStore []string
@@ -1183,194 +949,6 @@ func MustParseResolvedURLs(urlStrs []string) []*router.ResolvedURL {
 	return urls
 }
 
-func (s *StoreSuite) TestAddPromulgatedCharmDir(c *gc.C) {
-	charmDir := storetesting.Charms.CharmDir("wordpress")
-	s.checkAddCharm(c, charmDir, false, router.MustNewResolvedURL("~charmers/precise/wordpress-1", 1))
-}
-
-func (s *StoreSuite) TestAddPromulgatedCharmArchive(c *gc.C) {
-	charmArchive := storetesting.Charms.CharmArchive(c.MkDir(), "wordpress")
-	s.checkAddCharm(c, charmArchive, false, router.MustNewResolvedURL("~charmers/precise/wordpress-1", 1))
-}
-
-func (s *StoreSuite) TestAddUserOwnedCharmDir(c *gc.C) {
-	charmDir := storetesting.Charms.CharmDir("wordpress")
-	s.checkAddCharm(c, charmDir, false, router.MustNewResolvedURL("~charmers/precise/wordpress-1", -1))
-}
-
-func (s *StoreSuite) TestAddUserOwnedCharmArchive(c *gc.C) {
-	charmArchive := storetesting.Charms.CharmArchive(c.MkDir(), "wordpress")
-	s.checkAddCharm(c, charmArchive, false, router.MustNewResolvedURL("~charmers/precise/wordpress-1", -1))
-}
-
-func (s *StoreSuite) TestAddDevelopmentCharmArchive(c *gc.C) {
-	charmArchive := storetesting.Charms.CharmArchive(c.MkDir(), "wordpress")
-	url := router.MustNewResolvedURL("~charmers/development/precise/wordpress-1", 1)
-	s.checkAddCharm(c, charmArchive, false, url)
-}
-
-func (s *StoreSuite) TestAddBundleDir(c *gc.C) {
-	bundleDir := storetesting.Charms.BundleDir("wordpress-simple")
-	s.checkAddBundle(c, bundleDir, false, router.MustNewResolvedURL("~charmers/bundle/wordpress-simple-2", 3))
-}
-
-func (s *StoreSuite) TestAddBundleArchive(c *gc.C) {
-	bundleArchive, err := charm.ReadBundleArchive(
-		storetesting.Charms.BundleArchivePath(c.MkDir(), "wordpress-simple"),
-	)
-	s.addRequiredCharms(c, bundleArchive)
-	c.Assert(err, gc.IsNil)
-	s.checkAddBundle(c, bundleArchive, false, router.MustNewResolvedURL("~charmers/bundle/wordpress-simple-2", 3))
-}
-
-func (s *StoreSuite) TestAddUserOwnedBundleDir(c *gc.C) {
-	bundleDir := storetesting.Charms.BundleDir("wordpress-simple")
-	s.checkAddBundle(c, bundleDir, false, router.MustNewResolvedURL("~charmers/bundle/wordpress-simple-1", -1))
-}
-
-func (s *StoreSuite) TestAddUserOwnedBundleArchive(c *gc.C) {
-	bundleArchive, err := charm.ReadBundleArchive(
-		storetesting.Charms.BundleArchivePath(c.MkDir(), "wordpress-simple"),
-	)
-	c.Assert(err, gc.IsNil)
-	s.checkAddBundle(c, bundleArchive, false, router.MustNewResolvedURL("~charmers/bundle/wordpress-simple-1", -1))
-}
-
-func (s *StoreSuite) TestAddDevelopmentBundleArchive(c *gc.C) {
-	bundleArchive, err := charm.ReadBundleArchive(
-		storetesting.Charms.BundleArchivePath(c.MkDir(), "wordpress-simple"),
-	)
-	c.Assert(err, gc.IsNil)
-	url := router.MustNewResolvedURL("~charmers/development/bundle/wordpress-simple-2", 3)
-	s.checkAddBundle(c, bundleArchive, false, url)
-}
-
-func (s *StoreSuite) newStore(c *gc.C, withES bool) *Store {
-	var si *SearchIndex
-	if withES {
-		si = &SearchIndex{s.ES, s.TestIndex}
-	}
-	p, err := NewPool(s.Session.DB("juju_test"), si, &bakery.NewServiceParams{}, ServerParams{})
-	c.Assert(err, gc.IsNil)
-	store := p.Store()
-	p.Close()
-	return store
-}
-
-func (s *StoreSuite) TestAddCharmWithBundleSeries(c *gc.C) {
-	store := s.newStore(c, false)
-	defer store.Close()
-	ch := storetesting.Charms.CharmArchive(c.MkDir(), "wordpress")
-	err := store.AddCharmWithArchive(router.MustNewResolvedURL("~charmers/bundle/wordpress-2", -1), ch)
-	c.Assert(err, gc.ErrorMatches, `cannot read bundle archive: archive file "bundle.yaml" not found`)
-}
-
-func (s *StoreSuite) TestAddCharmWithMultiSeries(c *gc.C) {
-	store := s.newStore(c, false)
-	defer store.Close()
-	ch := storetesting.Charms.CharmArchive(c.MkDir(), "multi-series")
-	s.checkAddCharm(c, ch, false, router.MustNewResolvedURL("~charmers/multi-series-1", 1))
-	// Make sure it can be accessed with a number of names
-	e, err := store.FindEntity(router.MustNewResolvedURL("~charmers/multi-series-1", 1), nil)
-	c.Assert(err, gc.IsNil)
-	c.Assert(e.URL.String(), gc.Equals, "cs:~charmers/multi-series-1")
-	e, err = store.FindEntity(router.MustNewResolvedURL("~charmers/trusty/multi-series-1", 1), nil)
-	c.Assert(err, gc.IsNil)
-	c.Assert(e.URL.String(), gc.Equals, "cs:~charmers/multi-series-1")
-	e, err = store.FindEntity(router.MustNewResolvedURL("~charmers/wily/multi-series-1", 1), nil)
-	c.Assert(err, gc.IsNil)
-	c.Assert(e.URL.String(), gc.Equals, "cs:~charmers/multi-series-1")
-	_, err = store.FindEntity(router.MustNewResolvedURL("~charmers/precise/multi-series-1", 1), nil)
-	c.Assert(err, gc.ErrorMatches, "entity not found")
-	c.Assert(errgo.Cause(err), gc.Equals, params.ErrNotFound)
-}
-
-func (s *StoreSuite) TestAddCharmWithSeriesWhenThereIsAnExistingMultiSeriesVersion(c *gc.C) {
-	store := s.newStore(c, false)
-	defer store.Close()
-	ch := storetesting.Charms.CharmArchive(c.MkDir(), "multi-series")
-	err := store.AddCharmWithArchive(router.MustNewResolvedURL("~charmers/multi-series-1", -1), ch)
-	c.Assert(err, gc.IsNil)
-	ch = storetesting.Charms.CharmArchive(c.MkDir(), "wordpress")
-	err = store.AddCharmWithArchive(router.MustNewResolvedURL("~charmers/trusty/multi-series-2", -1), ch)
-	c.Assert(err, gc.ErrorMatches, `charm name duplicates multi-series charm name cs:~charmers/multi-series-1`)
-}
-
-func (s *StoreSuite) TestAddCharmWithMultiSeriesToES(c *gc.C) {
-	store := s.newStore(c, true)
-	defer store.Close()
-	ch := storetesting.Charms.CharmArchive(c.MkDir(), "multi-series")
-	s.checkAddCharm(c, ch, true, router.MustNewResolvedURL("~charmers/juju-gui-1", 1))
-}
-
-var addInvalidCharmURLTests = []string{
-	"cs:precise/wordpress-2",         // no user
-	"cs:~charmers/precise/wordpress", // no revision
-}
-
-func (s *StoreSuite) TestAddInvalidCharmURL(c *gc.C) {
-	store := s.newStore(c, false)
-	defer store.Close()
-	ch := storetesting.Charms.CharmArchive(c.MkDir(), "wordpress")
-	for i, urlStr := range addInvalidCharmURLTests {
-		c.Logf("test %d: %s", i, urlStr)
-		err := store.AddCharmWithArchive(&router.ResolvedURL{
-			URL:                 *charm.MustParseURL(urlStr),
-			PromulgatedRevision: -1,
-		}, ch,
-		)
-		c.Assert(err, gc.ErrorMatches, `charm added with invalid id .*`)
-	}
-}
-
-var addInvalidBundleURLTests = []string{
-	"cs:bundle/wordpress-2",         // no user
-	"cs:~charmers/bundle/wordpress", // no revision
-}
-
-func (s *StoreSuite) TestAddInvalidBundleURL(c *gc.C) {
-	store := s.newStore(c, false)
-	defer store.Close()
-	b := storetesting.Charms.BundleDir("wordpress-simple")
-	s.addRequiredCharms(c, b)
-	for i, urlStr := range addInvalidBundleURLTests {
-		c.Logf("test %d: %s", i, urlStr)
-		err := store.AddBundleWithArchive(&router.ResolvedURL{
-			URL:                 *charm.MustParseURL(urlStr),
-			PromulgatedRevision: -1,
-		}, b,
-		)
-		c.Assert(err, gc.ErrorMatches, `bundle added with invalid id .*`)
-	}
-}
-
-func (s *StoreSuite) TestAddBundleDuplicatingCharm(c *gc.C) {
-	store := s.newStore(c, false)
-	defer store.Close()
-	ch := storetesting.Charms.CharmDir("wordpress")
-	err := store.AddCharmWithArchive(router.MustNewResolvedURL("~charmers/precise/wordpress-2", -1), ch)
-	c.Assert(err, gc.IsNil)
-
-	b := storetesting.Charms.BundleDir("wordpress-simple")
-	s.addRequiredCharms(c, b)
-	err = store.AddBundleWithArchive(router.MustNewResolvedURL("~charmers/bundle/wordpress-5", -1), b)
-	c.Assert(err, gc.ErrorMatches, "bundle name duplicates charm name cs:~charmers/precise/wordpress-2")
-}
-
-func (s *StoreSuite) TestAddCharmDuplicatingBundle(c *gc.C) {
-	store := s.newStore(c, false)
-	defer store.Close()
-
-	b := storetesting.Charms.BundleDir("wordpress-simple")
-	s.addRequiredCharms(c, b)
-	err := store.AddBundleWithArchive(router.MustNewResolvedURL("~charmers/bundle/wordpress-simple-2", -1), b)
-	c.Assert(err, gc.IsNil)
-
-	ch := storetesting.Charms.CharmDir("wordpress")
-	err = store.AddCharmWithArchive(router.MustNewResolvedURL("~charmers/precise/wordpress-simple-5", -1), ch)
-	c.Assert(err, gc.ErrorMatches, "charm name duplicates bundle name cs:~charmers/bundle/wordpress-simple-2")
-}
-
 func (s *StoreSuite) TestOpenBlob(c *gc.C) {
 	charmArchive := storetesting.Charms.CharmArchive(c.MkDir(), "wordpress")
 	store := s.newStore(c, false)
@@ -1632,17 +1210,6 @@ func (s *StoreSuite) TestOpenCachedBlobFileWithFoundContent(c *gc.C) {
 	c.Assert(string(data), gc.Equals, expectContent)
 }
 
-func (s *StoreSuite) TestAddCharmWithUser(c *gc.C) {
-	store := s.newStore(c, false)
-	defer store.Close()
-
-	wordpress := storetesting.Charms.CharmDir("wordpress")
-	url := router.MustNewResolvedURL("cs:~who/precise/wordpress-23", -1)
-	err := store.AddCharmWithArchive(url, wordpress)
-	c.Assert(err, gc.IsNil)
-	assertBaseEntity(c, store, mongodoc.BaseURL(&url.URL), false)
-}
-
 func (s *StoreSuite) TestOpenCachedBlobFileWithNotFoundContent(c *gc.C) {
 	store := s.newStore(c, false)
 	defer store.Close()
@@ -1745,53 +1312,6 @@ func (s *StoreSuite) TestSESPutDoesNotErrorWithNoESConfigured(c *gc.C) {
 	defer store.Close()
 	err := store.UpdateSearch(nil)
 	c.Assert(err, gc.IsNil)
-}
-
-func (s *StoreSuite) TestAddCharmDirIndexed(c *gc.C) {
-	charmDir := storetesting.Charms.CharmDir("wordpress")
-	s.checkAddCharm(c, charmDir, true, router.MustNewResolvedURL("cs:~charmers/precise/wordpress-2", -1))
-}
-
-func (s *StoreSuite) TestAddCharmArchiveIndexed(c *gc.C) {
-	charmArchive := storetesting.Charms.CharmArchive(c.MkDir(), "wordpress")
-	s.checkAddCharm(c, charmArchive, true, router.MustNewResolvedURL("cs:~charmers/precise/wordpress-2", -1))
-}
-
-func (s *StoreSuite) TestAddBundleDirIndexed(c *gc.C) {
-	bundleDir := storetesting.Charms.BundleDir("wordpress-simple")
-	s.checkAddBundle(c, bundleDir, true, router.MustNewResolvedURL("cs:~charmers/bundle/baboom-2", -1))
-}
-
-func (s *StoreSuite) TestAddBundleArchiveIndexed(c *gc.C) {
-	bundleArchive, err := charm.ReadBundleArchive(
-		storetesting.Charms.BundleArchivePath(c.MkDir(), "wordpress-simple"),
-	)
-	c.Assert(err, gc.IsNil)
-	s.addRequiredCharms(c, bundleArchive)
-	s.checkAddBundle(c, bundleArchive, true, router.MustNewResolvedURL("cs:~charmers/bundle/baboom-2", -1))
-}
-
-func (s *StoreSuite) TestAddCharmDirIndexedAndPromulgated(c *gc.C) {
-	charmDir := storetesting.Charms.CharmDir("wordpress")
-	s.checkAddCharm(c, charmDir, true, router.MustNewResolvedURL("cs:~charmers/precise/wordpress-2", -1))
-}
-
-func (s *StoreSuite) TestAddCharmArchiveIndexedAndPromulgated(c *gc.C) {
-	charmArchive := storetesting.Charms.CharmArchive(c.MkDir(), "wordpress")
-	s.checkAddCharm(c, charmArchive, true, router.MustNewResolvedURL("cs:~charmers/precise/wordpress-2", 2))
-}
-
-func (s *StoreSuite) TestAddBundleDirIndexedAndPromulgated(c *gc.C) {
-	bundleDir := storetesting.Charms.BundleDir("wordpress-simple")
-	s.checkAddBundle(c, bundleDir, true, router.MustNewResolvedURL("cs:~charmers/bundle/baboom-2", 2))
-}
-
-func (s *StoreSuite) TestAddBundleArchiveIndexedAndPromulgated(c *gc.C) {
-	bundleArchive, err := charm.ReadBundleArchive(
-		storetesting.Charms.BundleArchivePath(c.MkDir(), "wordpress-simple"),
-	)
-	c.Assert(err, gc.IsNil)
-	s.checkAddBundle(c, bundleArchive, true, router.MustNewResolvedURL("cs:~charmers/bundle/baboom-2", 2))
 }
 
 var findBestEntityTests = []struct {
